🚨 Fix: 함수명변경, 위치 변경경
